+++ libs/dot-components/src/lib/components/avatar/Avatar.styles.ts
@@ -1,5 +1,6 @@
 import { Avatar } from '@material-ui/core';
 import styled, { css } from 'styled-components';
+import { AvatarProps } from './Avatar';
 
 export const rootClassName = 'dot-avatar';
 
@@ -9,11 +10,23 @@ const avatarSpacing = {
   large: 7,
 };
 
-export const StyledAvatar = styled(Avatar)`
+export const StyledAvatar = styled(Avatar)<AvatarProps>`
   ${({ theme }) => css`
     &.MuiAvatar-root {
-      background-color: ${theme.palette.grey[100]};
-      color: ${theme.palette.text.primary};
+      background-color: ${({ color }: AvatarProps) => {
+        return color && theme.palette.avatarColors[color]
+          ? theme.palette.avatarColors[color].backgroundColor
+          : theme.palette.avatarColors['default'].backgroundColor;
+      }};
+
+      .dot-i,
+      .dot-typography {
+        color: ${({ color }: AvatarProps) => {
+          return color && theme.palette.avatarColors[color]
+            ? theme.palette.avatarColors[color].color
+            : theme.palette.avatarColors['default'].color;
+        }};
+      }
 
       &.small {
         height: ${theme.spacing(avatarSpacing.small)}px;